% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/an_ma_bern.R
\name{an_ma_bern}
\alias{an_ma_bern}
\title{Analyse results from a fixed-sample multi-arm clinical trial for a Bernoulli
distributed primary outcome}
\usage{
an_ma_bern(des = des_ma_bern(), pval = rep(0.5 * des$alpha, des$K),
  alpha = des$alpha, correction = des$correction, sigmas,
  summary = F)
}
\arguments{
\item{des}{A \code{\link{list}} of class \code{"multiarm_des_ma_bern"}, as
returned by \code{\link{build_ma_bern}} or \code{\link{des_ma_bern}} (i.e., a
fixed-sample multi-arm clinical trial design for a Bernoulli distributed
outcome). Defaults to \code{des_ma_bern()}.}

\item{pval}{A \code{\link{matrix}} (or \code{\link{numeric}} vector that can
be coerced in to \code{\link{matrix}} form) whose rows indicate the values of
\ifelse{html}{\out{<b><i>p</i></b>}}{\eqn{\bold{p}}} for which to determine
which null hypotheses would be rejected. Defaults to
\code{rep(0.5*des$alpha, des$K)}.}

\item{alpha}{A \code{\link{numeric}} indicating the chosen value for
\ifelse{html}{\out{<i>&alpha;</i>}}{\eqn{\alpha}}, the significance level
(i.e., \code{alpha} can be used to modify/update \code{des} if desired).
Defaults to \code{des$alpha}.}

\item{correction}{A \code{\link{character}} string indicating the chosen
multiple comparison correction. Can be any of \code{"benjamini_hochberg"},
\code{"benjamini_yekutieli"}, \code{"bonferroni"}, \code{"dunnett"},
\code{"hochberg"}, \code{"holm_bonferroni"}, \code{"holm_sidak"},
\code{"none"}, \code{"sidak"}, and \code{"step_down_dunnett"}. Defaults to
\code{"des$correction"} (i.e., \code{correction} can be used to modify/update
\code{des} if desired).}

\item{summary}{A \code{\link{logical}} variable indicating whether a summary
of the function's progress should be printed to the console. Defaults to
\code{F}.}
}
\value{
A \code{\link{list}} containing the following elements
\itemize{
\item A \code{\link{tibble}} in the slot \code{$an} detailing which null
hypotheses would be rejected for each value of
\ifelse{html}{\out{<b><i>p</i></b>}}{\eqn{\bold{p}}}.
\item Each of the input variables, subject to possible internal modification.
}
}
\description{
\code{an_ma_bern()} analyses results from a specified fixed-sample multi-arm
clinical trial design for a Bernoulli distributed primary outcome variable,
to determine which null hypotheses to reject.
}
\examples{
# The results for the default parameters
an         <- an_ma_bern()
# An A-optimal design and a selection of possible p-values.
des_A      <- des_ma_bern(ratio = "A")
an_A       <- an_ma_bern(des_A, rbind(c(0.01, 0.01),
                                      c(0.025, 0.025),
                                      c(0.04, 0.04),
                                      c(0.05, 0.05),
                                      c(0.01, 0.06)))
# Using the root-K allocation rule, modifying the desired type of power and
# chosen multiple comparison correction, and a selection of possible
# p-values
des_root_K <- des_ma_bern(ratio      = rep(1/sqrt(2), 2),
                          correction = "holm_bonferroni",
                          power      = "disjunctive")
an_root_K  <- an_ma_bern(des_root_K, rbind(c(0.01, 0.01),
                                           c(0.025, 0.025),
                                           c(0.04, 0.04),
                                           c(0.05, 0.05),
                                           c(0.01, 0.06)))
}
\seealso{
\code{\link{build_ma_bern}}, \code{\link{des_ma_bern}},
\code{\link{opchar_ma_bern}}, \code{\link{plot.multiarm_des_ma_bern}},
\code{\link{sim_ma_bern}}.
}
