% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/des_dtl_bern.R
\name{des_dtl_bern}
\alias{des_dtl_bern}
\title{Design a multi-stage drop-the-losers multi-arm clinical trial for a Bernoulli
distributed primary outcome}
\usage{
des_dtl_bern(
  Kv = c(2, 1),
  alpha = 0.025,
  beta = 0.1,
  pi0 = 0.3,
  delta1 = 0.2,
  delta0 = 0,
  ratio = 1,
  power = "marginal",
  type = "variable",
  integer = F,
  summary = F
)
}
\arguments{
\item{Kv}{A \code{\link{numeric}} \code{\link{vector}} of strictly decreasing
values, indicating the chosen value for
\ifelse{html}{\out{<b><i>K</i></b>}}{\eqn{\bold{K}}}, the number of
experimental treatment arms present in each stage. Defaults to
\code{c(2, 1)}.}

\item{alpha}{A \code{\link{numeric}} indicating the chosen value for
\ifelse{html}{\out{<i>&alpha;</i>}}{\eqn{\alpha}}, the significance level
(family-wise error-rate). Defaults to \code{0.025}.}

\item{beta}{A \code{\link{numeric}} indicating the chosen value for
\ifelse{html}{\out{<i>&beta;</i>}}{\eqn{\beta}}, used in the definition of
the desired power. Defaults to \code{0.1}.}

\item{pi0}{A \code{\link{numeric}} indicating the chosen value for
\ifelse{html}{\out{<i>&pi;</i><sub>0</sub>}}{\eqn{\pi_0}}, the
response rate in the control arm. Defaults to \code{0.3}.}

\item{delta1}{A \code{\link{numeric}} indicating the chosen value for
\ifelse{html}{\out{<i>&delta;</i><sub>1</sub>}}{\eqn{\delta_1}}, the
'interesting' treatment effect. Defaults to \code{0.2}.}

\item{delta0}{A \code{\link{numeric}} indicating the chosen value for
\ifelse{html}{\out{<i>&delta;</i><sub>0</sub>}}{\eqn{\delta_0}}, the
'uninteresting' treatment effect. Defaults to \code{0}.}

\item{ratio}{A \code{\link{numeric}} indicating the chosen value for
\ifelse{html}{\out{<i>r</i>}}{\eqn{r}}, the stage-wise allocation ratio to
present experimental arms. Defaults to \code{1}.}

\item{power}{A \code{\link{character}} string indicating the chosen type of
power to design the trial for. Can be \code{"disjunctive"} or
\code{"marginal"}. Defaults to \code{"marginal"}.}

\item{type}{A \code{\link{character}} string indicating the choice for the
stage-wise sample size. Can be \code{"variable"} or \code{"fixed"}. Defaults
to \code{"variable"}.}

\item{integer}{A \code{\link{logical}} variable indicating whether the
computed possible sample sizes required in each arm in each stage should be
forced to be whole numbers. Defaults to \code{FALSE}. WARNING: If you set
\code{integer = TRUE} and \code{ratio != 1}, obscure results can occur due to
difficulties in identifying a suitable whole number sample size that meets
the allocation ratio requirement.}

\item{summary}{A \code{\link{logical}} variable indicating whether a summary
of the function's progress should be printed to the console. Defaults to
\code{FALSE}.}
}
\value{
A \code{\link{list}}, with additional class
\code{"multiarm_des_dtl_bern"}, containing the following elements:
\itemize{
\item A \code{\link{tibble}} in the slot \code{$opchar} summarising the
operating characteristics of the identified design.
\item A \code{\link{numeric}} in the slot \code{$e} specifying
\ifelse{html}{\out{<i>e</i>}}{\eqn{e}}, the trial's critical rejection
boundary for the final analysis.
\item A \code{\link{numeric}} in the slot \code{$maxN} specifying
\ifelse{html}{\out{<i>N</i>}}{\eqn{N}}, the trial's total required sample
size.
\item A \code{\link{numeric}} in the slot \code{$n_factor}, for internal use
in other functions.
\item A \code{\link{numeric}} in the slot \code{$n1} specifying
\ifelse{html}{\out{<i>n</i><sub>1</sub>}}{\eqn{n_1}}, the total sample size
required in stage one of the trial.
\item A \code{\link{numeric}} in the slot \code{$n10} specifying
\ifelse{html}{\out{<i>n</i><sub>10</sub>}}{\eqn{n_{10}}}, the sample size
required in the control arm in stage one of the trial.
\item Each of the input variables.
}
}
\description{
\code{des_dtl_bern()} determines multi-stage drop-the-losers multi-arm
clinical trial designs assuming the primary outcome variable is Bernoulli
distributed. It computes required design components and returns information
on key operating characteristics.
}
\examples{
# The design for the default parameters
des <- des_dtl_bern()
}
\seealso{
\code{\link{build_dtl_bern}}, \code{\link{gui}},
\code{\link{opchar_dtl_bern}}, \code{\link{plot.multiarm_des_dtl_bern}},
\code{\link{sim_dtl_bern}}.
}
