% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_ma.R
\name{gui_ma}
\alias{gui_ma}
\title{Plot operating characteristics of a fixed-sample multi-arm clinical trial}
\usage{
gui_ma()
}
\description{
\code{gui_ma()} run an R Shiny web browser based graphical user interface for
running the functions \code{\link{des_ma}} and
\code{\link{plot.multiarm_des_ma}}.
}
\examples{
# Launch the GUI
\dontrun{
gui_ma()
}
}
\seealso{
\code{\link{an_ma}}, \code{\link{build_ma}}, \code{\link{des_ma}},
\code{\link{des_int_ma}}, \code{\link{opchar_ma}},
\code{\link{plot.multiarm_des_ma}}, \code{\link{sim_ma}}, and their
associated S3 \code{\link{print}} and \code{\link{summary}} generics.
}
