% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opchar_ma.R
\name{opchar_ma}
\alias{opchar_ma}
\title{Analytically determine the operating characteristics of a fixed-sample
multi-arm clinical trial}
\usage{
opchar_ma(des = des_ma(), tau, summary = F)
}
\arguments{
\item{des}{A \code{\link{list}} of class \code{"multiarm_des_ma"}, as
returned by \code{\link{build_ma}}, \code{\link{des_ma}}, or
\code{\link{des_int_ma}} (i.e., a fixed-sample multi-arm clinical trial
design). Defaults to \code{des_ma()}.}

\item{tau}{A \code{\link{matrix}} whose rows indicate values of
\ifelse{html}{\out{<b><i>&tau;</i></b>}}{\eqn{\bold{\tau}}} at which to
evaluate the operating characteristics. Defaults internally to the global
null, global alternative, and each of the least favourable configurations,
for the specified design \code{des}.}

\item{summary}{A \code{\link{logical}} variable indicating whether a summary
of the function's progress should be printed to the console. Defaults to
\code{F}.}
}
\value{
A \code{\link{list}} of class \code{"multiarm_opchar_ma"} containing
the following elements
\itemize{
\item A \code{\link{tibble}} in the slot \code{$opchar} giving the determined
operating characteristics.
\item Each of the input variables, subject to possible internal modification.
}
}
\description{
\code{opchar_ma()} determines the operating characteristics of a specified
fixed-sample multi-arm clinical trial design, for given values of the true
treatment effects, using multivariate normal integration.
}
\examples{
# The operating characteristics for the default parameters
opchar        <- opchar_ma()
# An A-optimal design
des_A         <- des_ma(ratio = "A")
opchar_A      <- opchar_ma(des_A)
# Using the root-K allocation rule, modifying the desired type of power and
# chosen multiple comparison correction, and specifying tau explicitly
des_root_K    <- des_ma(ratio      = rep(1/sqrt(2), 2),
                        correction = "holm",
                        power      = "disjunctive")
opchar_root_K <- opchar_ma(des_root_K, rbind(c(0, 0),
                                             c(0.5, 0.5),
                                             c(0.5, 0),
                                             c(0, 0.5)))
}
\seealso{
\code{\link{an_ma}}, \code{\link{build_ma}}, \code{\link{des_ma}},
\code{\link{des_int_ma}}, \code{\link{gui_ma}},
\code{\link{plot.multiarm_des_ma}}, \code{\link{sim_ma}}.
}
