% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.multiarm_des_ma.R
\name{plot.multiarm_des_ma}
\alias{plot.multiarm_des_ma}
\title{Plot operating characteristics of a fixed-sample multi-arm clinical trial}
\usage{
\method{plot}{multiarm_des_ma}(x = des_ma(), delta_min = -x$delta1,
  delta_max = 2 * x$delta1, delta = delta_max - delta_min,
  density = 100, output = F, summary = F, ...)
}
\arguments{
\item{x}{A \code{\link{list}} of class \code{"multiarm_des_ma"}, as
returned by \code{\link{build_ma}}, \code{\link{des_ma}}, or
\code{\link{des_int_ma}} (i.e., a fixed-sample multi-arm clinical trial
design). Defaults to \code{des_ma()}.}

\item{delta_min}{A \code{\link{numeric}} specifying the chosen minimum value
for the treatment effects to include on the produced plots. Defaults to
\code{-x$delta1}.}

\item{delta_max}{A \code{\link{numeric}} specifying the chosen maximum
value for the treatment effects to include on the produced plots. Defaults to
\code{2*x$delta1}.}

\item{delta}{A \code{\link{numeric}} specifying the chosen treatment effect
shift to use the 'shifted treatment effects plot'. Defaults to
\code{delta_max - delta_min}.}

\item{density}{A \code{\link{numeric}} variable indicating the number of
treatment effect scenarios to consider for each power curve. Increasing
\code{density} increases the smoothness of the produced plots, at the cost of
increased run time. Defaults to \code{100}.}

\item{output}{A \code{\link{logical}} variable indicating whether the
available outputs from the function (see below) should be returned. Defaults
to \code{F}.}

\item{summary}{A \code{\link{logical}} variable indicating whether a summary
of the function's progress should be printed to the console. Defaults to
\code{F}.}
}
\value{
If \code{output = T}, a list containing the following elements
\itemize{
\item A \code{\link{list}} in the slot \code{$plots} containing the produced
plots.
\item Each of the input variables, subject to possible internal modification.
}
}
\description{
\code{plot.multiarm_des_ma()} produces power curve plots for a specified
fixed-sample multi-arm clinical trial design.
}
\examples{
# The design for the default parameters
des        <- des_ma()
plot(des)
# An A-optimal design, returning the avaiable outputs from
# plot.multiarm_des_ma()
des_A      <- des_ma(ratio = "A")
plots      <- plot(des_A, output = T)
# Using the root-K allocation rule, modifying the desired type of power, and
# choosing an alternative multiple comparison correction
des_root_K <- des_ma(ratio      = rep(1/sqrt(2), 2),
                     correction = "holm",
                     power      = "disjunctive")
plot(des_root_K)
}
\seealso{
\code{\link{build_ma}}, \code{\link{des_ma}},
\code{\link{des_int_ma}}, \code{\link{gui_ma}}, \code{\link{opchar_ma}},
\code{\link{plot.multiarm_des_ma}}, \code{\link{sim_ma}}.
}
