% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.multiarm_des_ma_bern.R
\name{plot.multiarm_des_ma_bern}
\alias{plot.multiarm_des_ma_bern}
\title{Plot operating characteristics of a fixed-sample multi-arm clinical trial
for a Bernoulli distributed primary outcome}
\usage{
\method{plot}{multiarm_des_ma_bern}(x = des_ma_bern(),
  delta_min = -x$pi0 + 1e-06, delta_max = 1 - x$pi0 - 1e-06,
  delta = x$delta1 - x$delta0, density = 100, output = F,
  summary = F, ...)
}
\arguments{
\item{x}{A \code{\link{list}} of class \code{"multiarm_des_ma_bern"}, as
returned by \code{\link{build_ma_bern}} or \code{\link{des_ma_bern}} (i.e., a
fixed-sample multi-arm clinical trial design for a Bernoulli distributed
outcome). Defaults to \code{des_ma_bern()}.}

\item{delta_min}{A \code{\link{numeric}} specifying the chosen minimum value
for the treatment effects to include on the produced plots. Defaults to
\code{-x$pi0 + 1e-6}.}

\item{delta_max}{A \code{\link{numeric}} specifying the chosen maximum
value for the treatment effects to include on the produced plots. Defaults to
\code{1 - x$pi0 - 1e-6}.}

\item{delta}{A \code{\link{numeric}} specifying the chosen treatment effect
shift to use in the 'shifted treatment effects plot'. Defaults to
\code{x$delta1 - x$delta0}.}

\item{density}{A \code{\link{numeric}} variable indicating the number of
treatment effect scenarios to consider for each power curve. Increasing
\code{density} increases the smoothness of the produced plots, at the cost of
increased run time. Defaults to \code{100}.}

\item{output}{A \code{\link{logical}} variable indicating whether the
available outputs from the function (see below) should be returned. Defaults
to \code{F}.}

\item{summary}{A \code{\link{logical}} variable indicating whether a summary
of the function's progress should be printed to the console. Defaults to
\code{F}.}

\item{...}{Not currently used.}
}
\value{
If \code{output = T}, a list containing the following elements
\itemize{
\item A \code{\link{list}} in the slot \code{$plots} containing the produced
plots.
\item Each of the input variables, subject to possible internal modification.
}
}
\description{
\code{plot.multiarm_des_ma_bern()} produces power curve plots for a specified
fixed-sample multi-arm clinical trial design assuming the primary outcome is
Bernoulli distributed.
}
\examples{
\dontrun{
# The design for the default parameters
des        <- des_ma_bern()
plot(des)
# An A-optimal design, returning the avaiable outputs from
# plot.multiarm_des_ma_bern()
des_A      <- des_ma_bern(ratio = "A")
plots      <- plot(des_A, output = TRUE)
# Using the root-K allocation rule, modifying the desired type of power, and
# choosing an alternative multiple comparison correction
des_root_K <- des_ma_bern(ratio      = rep(1/sqrt(2), 2),
                          correction = "holm_bonferroni",
                          power      = "disjunctive")
plot(des_root_K)
}
}
\seealso{
\code{\link{build_ma_bern}}, \code{\link{des_ma_bern}},
\code{\link{gui}}, \code{\link{opchar_ma_bern}},
\code{\link{plot.multiarm_des_ma_bern}}, \code{\link{sim_ma_bern}}.
}
