% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui.R
\name{gui}
\alias{gui}
\title{Graphical user interface to multi-arm trial design determination}
\usage{
gui()
}
\description{
\code{gui()} run an R Shiny web browser based graphical user interface for
running the functions \code{\link{des_ma}}, \code{\link{des_ma_bern}},
\code{\link{plot.multiarm_des_ma}}, and
\code{\link{plot.multiarm_des_ma_bern}}.
}
\examples{
# Launch the GUI
\dontrun{
gui()
}
}
\seealso{
\code{\link{des_ma}}, \code{\link{des_ma_bern}},
\code{\link{plot.multiarm_des_ma}}, \code{\link{plot.multiarm_des_ma_bern}}.
}
